Rotatable Bond Search
Created a function that searches for all Rotatable Bonds in a molecule. It determines a bond is rotatable if it doesn't find a loop in the molecule (finds the first index of the bond again)

This is not a true rule for finding rotatable bonds and it also doesn't factor in double bonds yet. This will be a future upgrade